+++ src/main/java/net/thesquire/backroomsmod/BackroomsModClient.java
@@ -5,7 +5,7 @@ import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
 import net.fabricmc.fabric.api.client.rendering.v1.BlockEntityRendererRegistry;
 import net.minecraft.client.render.RenderLayer;
 import net.thesquire.backroomsmod.block.ModBlocks;
-import net.thesquire.backroomsmod.block.entity.ModBlockEntities;
+import net.thesquire.backroomsmod.block.ModBlockEntities;
 import reborncore.client.multiblock.MultiblockRenderer;
 
 public class BackroomsModClient implements ClientModInitializer {