+++ src/main/java/net/thesquire/backroomsmod/block/custom/ModBlockWithEntity.java
@@ -1,0 +1,30 @@
+package net.thesquire.backroomsmod.block.custom;
+
+import net.minecraft.block.BlockState;
+import net.minecraft.block.BlockWithEntity;
+import net.minecraft.block.entity.BlockEntity;
+import net.minecraft.block.entity.BlockEntityTicker;
+import net.minecraft.block.entity.BlockEntityType;
+import net.minecraft.util.math.BlockPos;
+import org.jetbrains.annotations.Nullable;
+
+// This class is purely meant to widen access to the BlockWithEntity::checkType method
+public class ModBlockWithEntity extends BlockWithEntity {
+
+    /**
+     * {@return the ticker if the given type and expected type are the same, or {@code null} if they are different}
+     */
+    @Nullable
+    public static <E extends BlockEntity, A extends BlockEntity> BlockEntityTicker<A> checkType(BlockEntityType<A> givenType, BlockEntityType<E> expectedType, BlockEntityTicker<? super E> ticker) {
+        return BlockWithEntity.checkType(givenType, expectedType, ticker);
+    }
+
+    ///////////////////////////////////////////////////////////////////////
+
+    protected ModBlockWithEntity(Settings settings) { super(settings); }
+
+    @Nullable
+    @Override
+    public BlockEntity createBlockEntity(BlockPos pos, BlockState state) { return null; }
+
+}