+++ src/main/java/net/thesquire/backroomsmod/block/custom/MountableFluorescentLightBlock.java
@@ -1,13 +1,16 @@
 package net.thesquire.backroomsmod.block.custom;
 
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.HorizontalFacingBlock;
-import net.minecraft.block.ShapeContext;
+import net.minecraft.block.*;
+import net.minecraft.block.entity.BlockEntity;
+import net.minecraft.block.entity.BlockEntityTicker;
+import net.minecraft.block.entity.BlockEntityType;
 import net.minecraft.block.enums.WallMountLocation;
 import net.minecraft.item.ItemPlacementContext;
+import net.minecraft.screen.NamedScreenHandlerFactory;
 import net.minecraft.state.StateManager;
+import net.minecraft.state.property.BooleanProperty;
 import net.minecraft.state.property.EnumProperty;
+import net.minecraft.state.property.IntProperty;
 import net.minecraft.state.property.Properties;
 import net.minecraft.util.function.BooleanBiFunction;
 import net.minecraft.util.math.BlockPos;
@@ -15,11 +18,17 @@ import net.minecraft.util.math.Direction;
 import net.minecraft.util.shape.VoxelShape;
 import net.minecraft.util.shape.VoxelShapes;
 import net.minecraft.world.BlockView;
+import net.minecraft.world.World;
+import net.thesquire.backroomsmod.block.ModBlockEntities;
+import net.thesquire.backroomsmod.block.ModBlockProperties;
+import net.thesquire.backroomsmod.block.entity.flickering.MountableFluorescentLightBlockEntity;
 import org.jetbrains.annotations.Nullable;
 
-public class MountableFluorescentLightBlock extends HorizontalFacingBlock {
+public class MountableFluorescentLightBlock extends HorizontalFacingBlock implements BlockEntityProvider {
 
     public static final EnumProperty<WallMountLocation> WALL_MOUNT_LOCATION = Properties.WALL_MOUNT_LOCATION;
+    public static final BooleanProperty FLICKERING = ModBlockProperties.FLICKERING;
+    public static final IntProperty LUMINANCE = ModBlockProperties.LUMINANCE;
 
     private static final VoxelShape SHAPE_FLOOR_NS = VoxelShapes.combineAndSimplify(
             Block.createCuboidShape(0, 0, 6, 16, 1, 10),
@@ -69,8 +78,17 @@ public class MountableFluorescentLightBlock extends HorizontalFacingBlock {
             BooleanBiFunction.OR
     );
 
+    public static int getLuminance(BlockState state) {
+        return state.get(LUMINANCE);
+    }
+
+    /////////////////////////////////////////////////////////////////////////////////
+
     public MountableFluorescentLightBlock(Settings settings) {
         super(settings);
+        setDefaultState(getDefaultState().with(WALL_MOUNT_LOCATION, WallMountLocation.WALL)
+                .with(FACING, Direction.NORTH).with(FLICKERING, false)
+                .with(LUMINANCE, MountableFluorescentLightBlockEntity.defaultLuminance));
     }
 
     @Nullable
@@ -103,7 +121,44 @@ public class MountableFluorescentLightBlock extends HorizontalFacingBlock {
 
     @Override
     protected void appendProperties(StateManager.Builder<Block, BlockState> builder) {
-        builder.add(FACING, WALL_MOUNT_LOCATION);
+        builder.add(FACING, WALL_MOUNT_LOCATION, FLICKERING, LUMINANCE);
+    }
+
+    @Nullable
+    @Override
+    public BlockEntity createBlockEntity(BlockPos pos, BlockState state) {
+        return new MountableFluorescentLightBlockEntity(pos, state);
+    }
+
+    @Nullable
+    @Override
+    public <T extends BlockEntity> BlockEntityTicker<T> getTicker(World world, BlockState state, BlockEntityType<T> type) {
+        return ModBlockWithEntity.checkType(type, ModBlockEntities.MOUNTABLE_FLUORESCENT_LIGHT, MountableFluorescentLightBlockEntity::tick);
+    }
+
+    @SuppressWarnings("deprecation")
+    @Override
+    public BlockRenderType getRenderType(BlockState state) {
+        return BlockRenderType.MODEL;
+    }
+
+    @SuppressWarnings("deprecation")
+    @Override
+    public boolean onSyncedBlockEvent(BlockState state, World world, BlockPos pos, int type, int data) {
+        super.onSyncedBlockEvent(state, world, pos, type, data);
+        BlockEntity blockEntity = world.getBlockEntity(pos);
+        if (blockEntity == null) {
+            return false;
+        }
+        return blockEntity.onSyncedBlockEvent(type, data);
+    }
+
+    @SuppressWarnings("deprecation")
+    @Override
+    @Nullable
+    public NamedScreenHandlerFactory createScreenHandlerFactory(BlockState state, World world, BlockPos pos) {
+        BlockEntity blockEntity = world.getBlockEntity(pos);
+        return blockEntity instanceof NamedScreenHandlerFactory ? (NamedScreenHandlerFactory) blockEntity : null;
     }
 
 }