+++ src/main/java/net/thesquire/backroomsmod/dimension/ModDimensionKeys.java
@@ -8,7 +8,7 @@ import net.thesquire.backroomsmod.BackroomsMod;
 
 public class ModDimensionKeys {
 
-    public static final RegistryKey<World> LEVEL_0 = RegistryKey.of(Registry.WORLD_KEY,
-            new Identifier(BackroomsMod.MOD_ID, "level_0"));
+    public static final RegistryKey<World> LEVEL_0 = RegistryKey.of(Registry.WORLD_KEY, new Identifier(BackroomsMod.MOD_ID, "level_0"));
+    public static final RegistryKey<World> LEVEL_1 = RegistryKey.of(Registry.WORLD_KEY, new Identifier(BackroomsMod.MOD_ID, "level_1"));
 
 }