+++ src/main/java/net/thesquire/backroomsmod/world/ModWorldGen.java
@@ -4,7 +4,7 @@ import net.thesquire.backroomsmod.world.gen.ModOreGeneration;
 
 public class ModWorldGen {
 
-    public static void generateModWorldGen() {
+    public static void addModWorldGen() {
         ModOreGeneration.generateOres();
     }